refactor(sidebar-pane): move focus repair beside loop transitions
Move the unchanged focus helpers and their tests to loop state. Share their existing scene through the test fixture module.
